==> scripts/VHG code/0.01_quatro.lagos.R <==
#QUATRO LAGOS
#master script for all figures and stats




#load suryey data 
  #1.01_load.survey.R




#profile calculations
#include nutrients, other stuff

#how to figure out top/mid/bottom vs whole profiles? need multiple spreadsheets and ID col or something



#nutrient figure
  #2.21_nutrients.figs.R












#paleo stats and figure
  #4.02_paleo.R




#longterm doc and secchi stats and figs
  #4.21_longterm.R






#chl profile figure
  #8.20

#re-do this one ^^
